conversion to Vuetify
