+++ vue/src/components/ModelRunDetail.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import TimeSlider from "./TimeSlider.vue";
 import type { ModelRun } from "../client";
-import { CheckBadgeIcon } from "@heroicons/vue/24/solid";
 
 const props = defineProps<{
   modelRun: ModelRun;
@@ -18,64 +17,60 @@ async function handleClick() {
 </script>
 
 <template>
-  <details
-    class="group relative rounded-lg border-2 border-gray-50 open:border-blue-600 hover:border-gray-200 open:hover:border-blue-600"
-    :open="props.open"
+  <v-card
+    outlined
+    class="my-3 modelRunCard"
+    :class="{selectedCard: props.open}"
   >
-    <summary
-      class="cursor-pointer select-none list-none bg-gray-50 p-2 group-open:bg-blue-600 group-hover:bg-gray-200 group-open:group-hover:bg-blue-600"
+    <v-card-text
       @click.prevent="handleClick"
     >
-      <div class="grid grid-cols-8">
-        <div class="col-span-8 group-open:text-white">
-          <div class="model-title">
-            {{ modelRun.title }}
-          </div>
+      <v-row dense>
+        <div class="model-title">
+          {{ modelRun.title }}
         </div>
-        <div
-          class="col-span-3 text-xs font-light text-gray-600 group-open:text-gray-100"
-        >
+      </v-row>
+      <v-row dense>
+        <div>
           region:
         </div>
-        <div
-          class="col-span-5 font-mono text-xs font-light group-open:text-white"
-        >
+        <div>
           {{ modelRun.region?.name || "(none)" }}
         </div>
-        <div
-          class="col-span-3 text-xs font-light text-gray-600 group-open:text-gray-100"
-        >
+      </v-row>
+      <v-row dense>
+        <div>
           last updated:
         </div>
-        <div class="col-span-5 text-xs font-light group-open:text-white">
+        <div>
           {{
             modelRun.timestamp === null
               ? "--"
               : new Date(modelRun.timestamp * 1000).toLocaleString()
           }}
         </div>
-        <div
-          class="col-span-3 text-xs font-light text-gray-600 group-open:text-gray-100"
-        >
+      </v-row>
+      <v-row dense>
+        <div>
           number of sites:
         </div>
-        <div class="col-span-5 text-xs font-light group-open:text-white">
+        <div>
           {{ modelRun.numsites }}
         </div>
-        <div
-          class="col-span-3 text-xs font-light text-gray-600 group-open:text-gray-100"
-        >
+      </v-row>
+      <v-row dense>
+        <div>
           average score:
         </div>
-        <div class="col-span-5 text-xs font-light group-open:text-white">
+        <div>
           {{ modelRun.score === null ? "--" : modelRun.score.toFixed(2) }}
         </div>
-        <div
-          class="col-span-3 text-xs font-light text-gray-600 group-open:text-gray-100"
-        >
+      </v-row>
+      <v-row dense>
+        <div>
           date coverage:
         </div>
-        <div class="col-span-5 text-xs font-light group-open:text-white">
+        <div>
           {{
             modelRun.timerange === null
               ? "--"
@@ -92,38 +87,41 @@ async function handleClick() {
               )}`
           }}
         </div>
-        <div class="col-span-8 pt-2" />
-        <div
-          class="tooltip tooltip-right col-span-1 h-5 w-5 text-xs text-lime-600 group-open:text-white"
-          data-tip="Ground Truth"
+      </v-row>
+      <v-row dense>
+        <v-icon
+          v-if="modelRun.performer.short_code === 'TE' && modelRun.score == 1"
+          color="rgb(37, 99, 235)"
         >
-          <CheckBadgeIcon
-            v-if="modelRun.performer.short_code === 'TE' && modelRun.score == 1"
-          />
-        </div>
+          mdi-check-decagram
+        </v-icon>
         <div
-          class="col-span-7 rounded text-right text-xs text-gray-50 group-open:text-blue-600"
+          class="float-right mt-1 rounded bg-gray-400 pl-1 pr-1 group-open:bg-gray-100"
         >
-          <div
-            class="float-right mt-1 rounded bg-gray-400 pl-1 pr-1 group-open:bg-gray-100"
-          >
-            {{ modelRun.performer.short_code }}
-          </div>
+          {{ modelRun.performer.short_code }}
         </div>
-      </div>
-    </summary>
-    <div class="grid grid-cols-4 p-2">
-      <div class="col-span-4">
-        <TimeSlider
-          :min="modelRun.timerange?.min || 0"
-          :max="modelRun.timerange?.max || 0"
-        />
-      </div>
-    </div>
-  </details>
+      </v-row>
+    </v-card-text>
+    <v-card-actions v-if="open">
+      <TimeSlider
+        :min="modelRun.timerange?.min || 0"
+        :max="modelRun.timerange?.max || 0"
+      />
+    </v-card-actions>
+  </v-card>
 </template>
 
 <style scoped>
+.modelRunCard{
+  border: 3px solid transparent;
+}
+.modelRunCard:hover {
+  cursor: pointer;
+  border: 3px solid blue;
+}
+.selectedCard{
+  background-color: lightblue;
+}
 .model-title {
   max-width: 250px;
   overflow: hidden;
