+++ vue/src/components/ModelRunList.vue
@@ -246,93 +246,121 @@ watch([() => props.filters.region, () => props.filters.performer], () => {
 </script>
 
 <template>
-  <div class="flex flex-row bg-gray-100">
-    <span
-      v-if="!loading && !loadingSatelliteTimestamps"
-      style="font-size: 0.75em"
-      class="badge-accent badge ml-2"
-    >{{ totalModelRuns }} {{ totalModelRuns > 1 ? "Runs" : "Run" }}</span>
-    <span
-      v-if="!loading && !loadingSatelliteTimestamps && hasSatelliteImages && !state.satellite.satelliteImagesOn && state.filters.region_id?.length"
-      style="font-size: 0.75em"
-      class="badge-secondary badge ml-2"
-    >{{ hasSatelliteImages }} {{ hasSatelliteImages > 1 ? "Image Timestamps" : "Image Timestamp" }}</span>
-    <span
-      v-else-if="!loading && !loadingSatelliteTimestamps && state.satellite.satelliteTimeStamp && state.satellite.satelliteImagesOn && !satelliteRegionTooLarge"
-      style="font-size: 0.75em"
-      class="badge-secondary badge ml-2"
-    >Satellite Time: {{ state.satellite.satelliteTimeStamp }} - {{ state.satellite.satelliteTimeSource }}</span>
-    <div
-      v-if="loading"
-      class="px-2"
-      style="width: 100%"
-    >
-      <b>ModelRun</b>
-      <progress class="progress progress-primary" />
-    </div>
-    <div
-      v-if="loadingSatelliteTimestamps"
-      class="px-2"
-      style="width: 100%"
-    >
-      <b>Satellite Timestamps</b>
-      <progress class="progress progress-primary" />
-    </div>
-  </div>
-  <div v-if="state.satellite.loadingSatelliteImages">
-    <progress class="progress h-1 progress-primary" />
-  </div>
-  <div
-    v-if="!loading && state.filters.region_id?.length && satelliteRegionTooLarge"
-    style="font-size: 0.75em"
-    class=""
-  >
-    <div class="alert alert-warning shadow-lg">
-      <div>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          class="stroke-current flex-shrink-0 h-6 w-6"
-          fill="none"
-          viewBox="0 0 24 24"
-        ><path
-          stroke-linecap="round"
-          stroke-linejoin="round"
-          stroke-width="2"
-          d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-        /></svg>
-        <span>    Region is Too Large to efficiently get Images: 
-        </span>
-        <div class="flex-none">
-          <button
-            class="btn btn-xs btn-primary"
-            @click="satelliteRegionTooLarge = false"
-          >
-            Dismiss
-          </button>
-        </div>
+  <div>
+    <v-row>
+      <v-chip
+        v-if="!loading && !loadingSatelliteTimestamps"
+        style="font-size: 0.75em"
+        label
+        density="compact"
+        color="green"
+        class="ml-2"
+      >
+        {{ totalModelRuns }} {{ totalModelRuns > 1 ? "Runs" : "Run" }}
+      </v-chip>
+      <v-chip
+        v-if="!loading && !loadingSatelliteTimestamps && hasSatelliteImages && !state.satellite.satelliteImagesOn && state.filters.region_id?.length"
+        style="font-size: 0.75em"
+        label
+        density="compact"
+        color="pink"
+        class="ml-2"
+      >
+        {{ hasSatelliteImages }} {{ hasSatelliteImages > 1 ? "Image Timestamps" : "Image Timestamp" }}
+      </v-chip>
+      <v-chip
+        v-else-if="!loading && !loadingSatelliteTimestamps && state.satellite.satelliteTimeStamp && state.satellite.satelliteImagesOn && !satelliteRegionTooLarge"
+        style="font-size: 0.75em"
+        label
+        density="compact"
+        color="pink"
+        class="ml-2"
+      >
+        Satellite Time: {{ state.satellite.satelliteTimeStamp }} - {{ state.satellite.satelliteTimeSource }}
+      </v-chip>
+      <div
+        v-if="loading"
+        class="px-2"
+        style="width: 100%"
+      >
+        <b>ModelRun</b>
+        <v-progress-linear indeterminate />
       </div>
+      <div
+        v-if="loadingSatelliteTimestamps"
+        class="px-2"
+        style="width: 100%"
+      >
+        <b>Satellite Timestamps</b>
+        <v-progress-linear indeterminate />
+      </div>
+    </v-row>
+    <div
+      v-if="state.satellite.loadingSatelliteImages"
+      class="mt-5"
+    >
+      <v-progress-linear
+        indeterminate
+        height="8"
+        color="primary"
+      />
     </div>
-  </div>
-  <div
-    class="flex flex-col gap-2 overflow-y-scroll p-2"
-    @scroll="handleScroll"
-  >
-    <ModelRunDetail
-      v-for="modelRun in modelRuns"
-      :key="modelRun.key"
-      :model-run="modelRun"
-      :open="openedModelRuns.has(modelRun.key)"
-      :class="{
-        outlined: hoveredInfo.region.includes(
-          `${modelRun.id}_${modelRun.region?.id}_${modelRun.performer.id}`
-        ),
-      }"
-      @toggle="() => handleToggle(modelRun)"
-    />
+    <div
+      v-if="!loading && state.filters.region_id?.length && satelliteRegionTooLarge"
+      class="mt-5"
+      style="font-size: 0.75em"
+    >
+      <v-alert type="warning">
+        <div>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            class="stroke-current flex-shrink-0 h-6 w-6"
+            fill="none"
+            viewBox="0 0 24 24"
+          ><path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            stroke-width="2"
+            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
+          /></svg>
+          <span>    Region is Too Large to efficiently get Images: 
+          </span>
+          <div class="flex-none">
+            <button
+              class="btn btn-xs btn-primary"
+              @click="satelliteRegionTooLarge = false"
+            >
+              Dismiss
+            </button>
+          </div>
+        </div>
+      </v-alert>
+    </div>
+    <v-container
+      class="overflow-y-auto p-5 mt-5 modelRuns"
+      @scroll="handleScroll"
+    >
+      <ModelRunDetail
+        v-for="modelRun in modelRuns"
+        :key="modelRun.key"
+        :model-run="modelRun"
+        :open="openedModelRuns.has(modelRun.key)"
+        :class="{
+          outlined: hoveredInfo.region.includes(
+            `${modelRun.id}_${modelRun.region?.id}_${modelRun.performer.id}`
+          ),
+        }"
+        @toggle="() => handleToggle(modelRun)"
+      />
+    </v-container>
   </div>
 </template>
 
 <style scoped>
+.modelRuns {
+  height: calc(100vh - 250px);
+
+}
 .outlined {
   background-color: orange;
   border: 3px solid orange;
