+++ vue/src/interactions/mouseEvents.ts
@@ -1,13 +1,33 @@
-import { Ref, ref } from "vue";
+import { App, Ref, createApp, defineComponent, nextTick, ref } from "vue";
 import { Color, Map, MapLayerMouseEvent, Popup } from "maplibre-gl";
 import { ShallowRef } from "vue";
 import { getSiteObservationDetails, selectedObservationList, state } from "../store";
+import PopUpComponent from '../components/PopUpComponent.vue'
 
+import 'vuetify/styles'
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+import '@mdi/font/css/materialdesignicons.css' // Ensure you are using css-loader
+
+const vuetify = createVuetify({
+  components,
+  directives,
+})
+
+interface PopUpData {
+  siteId: string;
+  siteColor: string;
+  score: number;
+  groundTruth: boolean;
+  scoreColor: string;
+}
 const checkBadge =
   '<svg style="display:inline;" width="20" height="20" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="replacementColor" aria-hidden="true"><path fill-rule="evenodd" d="M8.603 3.799A4.49 4.49 0 0112 2.25c1.357 0 2.573.6 3.397 1.549a4.49 4.49 0 013.498 1.307 4.491 4.491 0 011.307 3.497A4.49 4.49 0 0121.75 12a4.49 4.49 0 01-1.549 3.397 4.491 4.491 0 01-1.307 3.497 4.491 4.491 0 01-3.497 1.307A4.49 4.49 0 0112 21.75a4.49 4.49 0 01-3.397-1.549 4.49 4.49 0 01-3.498-1.306 4.491 4.491 0 01-1.307-3.498A4.49 4.49 0 012.25 12c0-1.357.6-2.573 1.549-3.397a4.49 4.49 0 011.307-3.497 4.49 4.49 0 013.497-1.307zm7.007 6.387a.75.75 0 10-1.22-.872l-3.236 4.53L9.53 12.22a.75.75 0 00-1.06 1.06l2.25 2.25a.75.75 0 001.14-.094l3.75-5.25z" clip-rule="evenodd"></path></svg>';
 
   const hoveredInfo: Ref<{region: string[], siteId: number[]}> = ref({region: [], siteId:[]});
 
+let app: App | null = null;
 const calculateScoreColor = (score: number) => {
   if (score <= 0.25) {
     return "red";
@@ -28,7 +48,7 @@ const popupLogic = (map: ShallowRef<null | Map>) => {
   const popup = new Popup({
     closeButton: false,
     closeOnClick: false,
-    maxWidth: '400px',
+    maxWidth: '550px',
   });
   let insideObservation = false;
   const drawPopup = (e: MapLayerMouseEvent) => {
@@ -40,6 +60,7 @@ const popupLogic = (map: ShallowRef<null | Map>) => {
       insideObservation = true;
       hoveredInfo.value.region = [];
       hoveredInfo.value.siteId = [];
+      const popupData: PopUpData[] = [];
       e.features.forEach(
         (
           item: GeoJSON.GeoJsonProperties & {
@@ -88,13 +109,35 @@ const popupLogic = (map: ShallowRef<null | Map>) => {
                     2
                     )}</div></li>`;
                   }
+                  popupData.push({
+                    siteId: `${regionName}_${String(id).padStart(4, '0')}`,
+                    score,
+                    groundTruth: item.properties.groundtruth,
+                    siteColor: `rgb(${fillColor.r *255}, ${fillColor.g * 255}, ${fillColor.b * 255})`,
+                    scoreColor: calculateScoreColor(score),
+                  })    
               }
             }
           }
         }
       );
       html += "</ul></div>";
-      popup.setLngLat(coordinates).setHTML(html).addTo(map.value);
+      popup.setLngLat(coordinates).setHTML('<div id="popup-content"></div>').addTo(map.value);
+      const MyNewPopup = defineComponent({
+        extends: PopUpComponent,
+        setup() {
+          return { data: popupData }
+        },
+      })
+      nextTick(() => {
+        if (app !== null) {
+          app.unmount();
+          app = null;
+        }
+        app = createApp(MyNewPopup).use(vuetify)
+        app.mount('#popup-content');
+      })
+      
     } else if (map.value) {
       hoveredInfo.value.region = [];
       hoveredInfo.value.siteId = [];
